+++ service/unittest/default_k8sclient.go
@@ -1,18 +1,18 @@
 package unittest
 
 import (
-	securityv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/security/v1alpha1"
-	"github.com/giantswarm/apiextensions/v3/pkg/clientset/versioned"
-	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
-	"github.com/giantswarm/k8sclient/v5/pkg/k8scrdclient"
+	"github.com/giantswarm/k8sclient/v6/pkg/k8sclient"
+	"github.com/giantswarm/k8sclient/v6/pkg/k8scrdclient"
 	v1 "k8s.io/api/core/v1"
-	apiextensionsclient "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
+	"k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake" // nolint:staticcheck
+
+	securityv1alpha1 "github.com/giantswarm/organization-operator/api/v1alpha1"
 )
 
 type fakeK8sClient struct {
@@ -36,7 +36,7 @@ func FakeK8sClient() k8sclient.Interface {
 		}
 
 		k8sClient = &fakeK8sClient{
-			ctrlClient: fake.NewFakeClientWithScheme(scheme),
+			ctrlClient: fake.NewClientBuilder().WithScheme(scheme).Build(),
 			scheme:     scheme,
 		}
 	}
@@ -56,11 +56,7 @@ func (f *fakeK8sClient) DynClient() dynamic.Interface {
 	return nil
 }
 
-func (f *fakeK8sClient) ExtClient() apiextensionsclient.Interface {
-	return nil
-}
-
-func (f *fakeK8sClient) G8sClient() versioned.Interface {
+func (f *fakeK8sClient) ExtClient() clientset.Interface {
 	return nil
 }